Modificación html de llamadas y corrección de errores

+++ src/main/java/ar/edu/unnoba/poo2021/Fiscalia/model/LlamadaTelefonica.java
@@ -27,7 +27,7 @@ public class LlamadaTelefonica extends Informacion {
         this.numeroReceptor = numeroReceptor;
     }
 
-    public long getDuracion() {
+    public Integer getDuracion() {
         return duracion;
     }
